moved the navigation bar in its own component.
