+++ appearance/src/components/navigation-bar.css
@@ -1,0 +1,9 @@
+
+.bg-myBlue {
+    background-color: RoyalBlue;
+}
+
+.mb {
+	margin-bottom: 35px;
+}
+
